requirements.txt
flet==0.22.*
Pyrebase4
mysql-connector-python
setuptools